requirements.txt
# Core
pydantic-settings>=2.2.0
structlog>=24.1.0
python-dotenv>=1.0.1

# Database
psycopg>=3.1.18
sqlalchemy>=2.0.29

# Cache / Broker
redis>=5.0.3

# Message Broker
confluent-kafka>=2.4.0

# Spark
pyspark>=3.5.0
py4j>=0.10.9.7

# Document Parsers
pdfplumber>=0.11.0
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.3
lxml>=5.1.0
html2text>=2024.2.26
chardet>=5.2.0

# OCR
pytesseract>=0.3.10
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80

# Async HTTP
aiohttp>=3.9.3

# Dev
black>=24.2.0
ruff>=0.3.0
mypy>=1.8.0